+++ crates/runner-host/src/board/crypto.rs
@@ -14,6 +14,14 @@
 
 use wasefire_board_api as board;
 
+use crate::board::Board;
+
 mod ccm;
 
-impl board::crypto::Api for crate::board::Board {}
+impl board::crypto::Api for &mut Board {
+    type Ccm<'a> = &'a mut Board
+    where Self: 'a;
+    fn ccm(&mut self) -> Self::Ccm<'_> {
+        self
+    }
+}