+++ crates/runner-nordic/src/tasks/crypto.rs
@@ -14,6 +14,13 @@
 
 use wasefire_board_api as board;
 
+use crate::tasks::Board;
+
 mod ccm;
 
-impl board::crypto::Api for crate::tasks::Board {}
+impl board::crypto::Api for &mut Board {
+    type Ccm<'a> = &'a mut Board where Self: 'a;
+    fn ccm(&mut self) -> Self::Ccm<'_> {
+        self
+    }
+}