+++ crates/scheduler/CHANGELOG.md
@@ -18,4 +18,4 @@
 
 ## 0.1.0
 
-<!-- Update PR number to skip CHANGELOG.md test: #112 -->
+<!-- Update PR number to skip CHANGELOG.md test: #113 -->